+++ src/app/components/lodash.extended.js
@@ -14,10 +14,6 @@ function () {
       array.splice(toIndex, 0, array.splice(fromIndex, 1)[0]);
       return array;
     },
-    remove: function (array, index) {
-      array.splice(index, 1);
-      return array;
-    },
     // If variable is value, then return alt. If variable is anything else, return value;
     toggle: function (variable, value, alt) {
       return variable === value ? alt : value;