+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.0.0 (unreleased)
+
+**Enhancements**
+- [Issue #1297](https://github.com/grafana/grafana/issues/1297). Graphite: Added cumulative and minimumBelow graphite functions
+- [Issue #1296](https://github.com/grafana/grafana/issues/1296). InfluxDB: Auto escape column names with special characters. Thanks @steven-aerts
+
 # 1.9.1 (2014-12-29)
 
 **Enhancements**